tenable_sc.vulnerability: Update mapping of `tenable_sc.vulnerability.plugin.text` field to match_only_text (#16508)
The "tenable_sc.vulnerability.plugin.text" field stores several 
details of vulnerable packages found on the host. It is currently 
mapped to "keyword". It can grow larger than maximum "ignore_above" 
limit of keywordfields, making it unusable in searches.

This PR updates the data type of "tenable_sc.vulnerability.plugin.text" 
field to "match_only_text" to allow searches on this field.

This is a breaking-change because the data types "keyword" and 
"match_only_text" are from different family [1] and could break 
users' custom Kibana dashboards or queries.

[1] https://www.elastic.co/docs/reference/elasticsearch/mapping-reference/field-data-types