carbon_black_cloud: fix pagination logic for asset_vulnerability_summary data stream
For CEL input, it correctly sets the `cursor.processed_num_rows` to 0 when it's the last page.

For HTTPJSON input, it updates the `body.rows` to 1000 and correctly updates the `body.start`
to fetch the next page. It also adds `do_not_log_failure: true` in set processors to avoid
updating fleet health status to degraded.

This change bumps minimum Kibana version to ^8.19.4 || ~9.0.7 || ^9.1.4