correct next_page value in system test rule config (#16661)
tines: correct `next_page` value in system test rule config to prevent updating
the fleet health status to degraded

Previously, the `next_page` value was incorrectly set to the string "null", so during system testing
it attempted to perform a new request with that URL (with the value `null`), resulting in an error.

Actually, it should be `"next_page": null` as mentioned in the (documentation)[1].

No changelog entry or new version because this shouldn't affect what's in the built package.

[1] https://www.tines.com/api/welcome/#the-meta-object